+++ packages/terra-framework-docs/src/terra-dev-site/test/date-time-picker/DateTimePickerInvalidIncomplete.test.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import DateTimePickerExampleTemplate from './common/DateTimePickerExampleTemplate';
+
+const DateTimePickerInvalidIncomplete = () => (
+  <DateTimePickerExampleTemplate
+    isIncomplete
+    required
+    isInvalid
+  />
+);
+
+export default DateTimePickerInvalidIncomplete;